requirements.txt
flask>=3.0.0
pandas>=2.2.0
openpyxl>=3.1.0
rapidfuzz>=3.9.0
requests>=2.32.0
pillow>=10.0.0
